Add support for string literals
