+++ src/shared/Minsk.Core/CodeAnalysis/Syntax/Lexer.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Runtime.InteropServices;
+using System.Text;
 using Minsk.Core.CodeAnalysis.Text;
 
 namespace Minsk.Core.CodeAnalysis.Syntax
@@ -21,6 +22,7 @@ namespace Minsk.Core.CodeAnalysis.Syntax
         public DiagnosticsBag Diagnostics { get; }
 
         private char Current => Peek(0);
+        private char LookAhead => Peek(1);
 
         private char Peek(int offset)
         {
@@ -167,6 +169,9 @@ namespace Minsk.Core.CodeAnalysis.Syntax
                 case '9':
                     ReadNumberToken();
                     break;
+                case '"':
+                    ReadString();
+                    break;
                 case ' ':
                 case '\t':
                 case '\n':
@@ -194,6 +199,47 @@ namespace Minsk.Core.CodeAnalysis.Syntax
             var text = _kind.GetText() ?? _text.ToString(_start, length);
 
             return new SyntaxToken(_kind, _start, text, _value);
+        }
+
+        private void ReadString()
+        {
+            // Skip the starting quote
+            _position++;
+            var sb = new StringBuilder();
+            var done = false;
+
+            while (!done)
+            {
+                switch (Current)
+                {
+                    case '\0':
+                    case '\r':
+                    case '\n':
+                        var span = new TextSpan(_start, 1);
+                        Diagnostics.ReportUnterminatedString(span);
+                        done = true;
+                        break;
+                    case '"':
+                        if (LookAhead == '"')
+                        {
+                            sb.Append(Current);
+                            _position += 2;
+                        }
+                        else
+                        {
+                            _position++;
+                            done = true;
+                        }
+                        break;
+                    default:
+                        sb.Append(Current);
+                        _position++;
+                        break;
+                }
+            }
+
+            _kind = SyntaxKind.StringToken;
+            _value = sb.ToString();
         }
 
         private void ReadWhiteSpace()
