+++ src/shared/Minsk.Core/CodeAnalysis/Syntax/SyntaxFacts.cs
@@ -42,6 +42,7 @@ namespace Minsk.Core.CodeAnalysis.Syntax
                 case SyntaxKind.PlusToken:
                 case SyntaxKind.MinusToken:
                 case SyntaxKind.BangToken:
+
                 case SyntaxKind.TildeToken:
                     return 6;
                 default:
