+++ src/shared/Minsk.Core/CodeAnalysis/Syntax/SyntaxTree.cs
@@ -35,20 +35,39 @@ namespace Minsk.Core.CodeAnalysis.Syntax
 
         public static IEnumerable<SyntaxToken> ParseTokens(string text)
         {
+            return ParseTokens(text, out _);
+        }
+
+        public static IEnumerable<SyntaxToken> ParseTokens(string text, out ImmutableArray<Diagnostic> diagnostics)
+        {
             var sourceText = SourceText.From(text);
-            return ParseTokens(sourceText);
+            return ParseTokens(sourceText, out diagnostics);
         }
 
         public static IEnumerable<SyntaxToken> ParseTokens(SourceText text)
         {
-            var lexer = new Lexer(text);
-            while (true)
-            {
-                var token = lexer.Lex();
-                if (token.Kind == SyntaxKind.EofToken) break;
+            return ParseTokens(text, out _);
+        }
 
-                yield return token;
+        public static IEnumerable<SyntaxToken> ParseTokens(SourceText text, out ImmutableArray<Diagnostic> diagnostics)
+        {
+            IEnumerable<SyntaxToken> LexTokens(Lexer lexer)
+            {
+                while (true)
+                {
+                    var token = lexer.Lex();
+                    if (token.Kind == SyntaxKind.EofToken)
+                        break;
+
+                    yield return token;
+                }
             }
+
+            var l = new Lexer(text);
+            var result = LexTokens(l).ToImmutableArray();
+            diagnostics = l.Diagnostics.ToImmutableArray();
+
+            return result;
         }
     }
 }
